+++ interfaces/irc/__init__.py
@@ -1,0 +1,7 @@
+from interfaces.irc.commander import IRCCommander, IRCCommand
+from interfaces.irc.ignore import IgnoreList, IgnoreEntryAlreadyExistsError
+from interfaces.irc.irc import IRC
+from interfaces.irc.logger import IRCChannelLogger, IRCQueryLogger, IRCLoggerSource
+from interfaces.irc.interface import NanoIRC
+from interfaces.irc.network import Network
+from interfaces.irc.postmaster import Postmaster
