--- src/irc/__init__.py
@@ -1,7 +1,0 @@
-from .commander import IRCCommander, IRCCommand
-from .ignore import IgnoreList, IgnoreEntryAlreadyExistsError
-from .irc import IRC
-from .logger import IRCChannelLogger, IRCQueryLogger, IRCLoggerSource
-from .nano_irc import NanoIRC
-from .network import Network
-from .postmaster import Postmaster
